Move code to more appropriate files.
Partial #533
